Migrating ContentRepositoryExtensions - combining it with existing file in Umbraco.Core
